pyproject.toml
[project]
name = "claude-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.9"
dependencies = [
    "aiohttp>=3.13.3",
    "beautifulsoup4>=4.14.3",
    "claude-agent-sdk>=0.1.18",
    "dataclasses-json>=0.6.7",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "html2text>=2025.4.15",
    "notion-client>=2.7.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
]
